Updated timeslot to allow blank end time
Previously the user was required to enter a start and end time for each
timeslot.  Now users can clear the end time when not needed / known.

Temporary change to displaying the event timeline where timeslots
without end times are set to 10 pm will be removed (was used to allow
clicking on timeslot for testing)